refactor: repository 인터페이스를 패키지로 묶음
