+++ src/main/java/roomescape/domain/repository/MemberRepository.java
@@ -1,7 +1,8 @@
-package roomescape.domain;
+package roomescape.domain.repository;
 
 import java.util.List;
 import java.util.Optional;
+import roomescape.domain.Member;
 
 public interface MemberRepository {
